Implement BO_ADVANCE_PROTECTION skill

+++ src/map/skill.h
@@ -2627,8 +2627,8 @@ struct skill_interface {
 	int (*additional_effect) (struct block_list* src, struct block_list *bl, uint16 skill_id, uint16 skill_lv, int attack_type, int dmg_lv, int64 tick);
 	int (*counter_additional_effect) (struct block_list* src, struct block_list *bl, uint16 skill_id, uint16 skill_lv, int attack_type, int64 tick);
 	int (*blown) (struct block_list* src, struct block_list* target, int count, enum unit_dir dir, int flag);
-	int (*break_equip) (struct block_list *bl, unsigned short where, int rate, int flag);
-	int (*strip_equip) (struct block_list *bl, unsigned short where, int rate, int lv, int time);
+	int (*break_equip) (struct block_list *bl, uint32 where, int rate, int flag);
+	int (*strip_equip) (struct block_list *bl, uint32 where, int rate, int lv, int time);
 	struct skill_unit_group* (*id2group) (int group_id);
 	struct skill_unit_group *(*unitsetting) (struct block_list* src, uint16 skill_id, uint16 skill_lv, short x, short y, int flag);
 	struct skill_unit *(*initunit) (struct skill_unit_group *group, int idx, int x, int y, int val1, int val2);
